// src/edgeiq-os/services/performance-intelligence/types.ts
export type HistoricalPerformanceIntelligenceRow = {
  current_race_context_key: string;
  runner_number: string;
  performance_fact_id: string;
  legacy_performance_fact_id: string;
  race_context_key: string;
  race_date: string;
  track: string;
  state: string;
  race_number: string;
  race_name: string;
  race_class: string;
  distance_metres: string;
  track_condition: string;
  track_rating: string;
  rail_position: string;
  horse: string;
  horse_code: string;
  trainer: string;
  jockey: string;
  barrier: string;
  weight: string;
  finish_position: string;
  margin_raw: string;
  margin_lengths_raw: string;
  starting_price: string;
  governed_race_time_seconds: string;
  performance_quality_state: string;
  benchmark_exclusion_reason: string;
  source_snapshot_ids?: string[];
  selected_benchmark_id: string;
  benchmark_level: string;
  benchmark_confidence: string;
  benchmark_sample_size: string;
  benchmark_time_seconds: string;
  race_seconds_vs_benchmark: string;
  fallback_path: string;
  fingerprint_pattern: string;
  fingerprint_quality_state: string;
};